Make a factory class for bot
